Route herdr agent state through pane monitors and add detach
herdr's agent reports now drive the pane monitor directly, adopting the
manifest agent by herdr's id and committing state without a settle
window, so badges, cards, rollups, Live Activity counts, and
notifications follow herdr in both control-mode variants. Screen
classification and title identity stand down while herdr names an
agent, and a pane missing from a snapshot's agent list is cleared.
herdr's pane directories feed project identity through the gateway's
host key.

Detach from herdr from any gateway or projected tab's context menu,
optionally closing the gateway, and projected-tab closes honor the tmux
close-action setting. Also: divider drags resize through herdr, the
gateway can auto-hide on attach, a health ping and writer overflow
re-snapshot panes, discovery reports control-stream support for the
picker, projected tabs get an H badge, and herdr families stay put
during window transfers.

+++ rootshell/Features/Herdr/HerdrControlProtocol.swift
@@ -100,7 +100,8 @@ nonisolated enum HerdrControl {
         let pane_id: String
         /// "left", "right", "up", or "down".
         let direction: String
-        let amount: Int
+        /// Fraction of the split to move.
+        let amount: Double
     }
 
     // MARK: - Responses and records
